[ADC] Regenerated demo for adc_adc_window_sleepwalking_sam_e54_xpro
MH3-79182

+++ apps/adc/adc_window_sleepwalking/firmware/src/config/sam_e54_xpro/peripheral/rtc/plib_rtc_timer.c
@@ -44,7 +44,6 @@
 #include "interrupts.h"
 #include "plib_rtc.h"
 #include <stdlib.h>
-#include <limits.h>
 
 
 void RTC_Initialize(void)
@@ -84,12 +83,12 @@ void RTC_Initialize(void)
 bool RTC_PeriodicIntervalHasCompleted (RTC_PERIODIC_INT_MASK period)
 {
    bool periodIntervalComplete = false;
-   if( (RTC_REGS->MODE0.RTC_INTFLAG & period) == period)
+   if( (RTC_REGS->MODE0.RTC_INTFLAG & (uint16_t)period) == (uint16_t)period)
    {
        periodIntervalComplete = true;
 
        /* Clear Periodic Interval Interrupt */
-       RTC_REGS->MODE0.RTC_INTFLAG = period;
+       RTC_REGS->MODE0.RTC_INTFLAG = (uint16_t)period;
    }
 
     return periodIntervalComplete;
@@ -144,7 +143,7 @@ void RTC_Timer32CountSyncEnable ( void )
         /* Wait for Synchronization */
     }
 }
-	
+
 void RTC_Timer32CountSyncDisable ( void )
 {
     RTC_REGS->MODE0.RTC_CTRLA &= (uint16_t)(~RTC_MODE0_CTRLA_COUNTSYNC_Msk);
@@ -154,7 +153,7 @@ void RTC_Timer32CountSyncDisable ( void )
         /* Wait for Synchronization */
     }
 }
-	
+
 void RTC_Timer32Start ( void )
 {
     RTC_REGS->MODE0.RTC_CTRLA |= RTC_MODE0_CTRLA_ENABLE_Msk;
